requirements.txt
django
djangorestframework
pillow
psycopg2-binary
python-decouple
drf-yasg